modified int value to double value

+++ backend/excel-school-app/DTOs/Grade/UpdateGradeDto.cs
@@ -5,7 +5,7 @@ namespace excel_school_app.DTOs.Grade
     public class UpdateGradeDto
     {
         public Enums.Subject Subject { get; set; }
-        public int Value { get; set; }
+        public double Value { get; set; }
         public string Appreciation { get; set; } = string.Empty;
         public DateOnly ExamDate { get; set; }
         public Enums.Term Term { get; set; }
